Move old sodetlib code to legacy dir
